requirements.txt
python-dotenv>=1.0.0
transformers>=4.40.0
peft>=0.10.0
torch>=2.0.0
accelerate>=0.30.0
